requirements.txt
requests>=2.31.0
urllib3>=2.0.0
google-genai>=1.5.0
